backend/foodgram/requirements.txt
Django==3.2
djangorestframework==3.12.4
psycopg2-binary==2.9.5
djoser==2.1.0
djangorestframework-simplejwt==4.8.0
Pillow==9.5.0